// shesha-reactjs/src/apis/permission.tsx
export interface PermissionDto {
  id?: string | null;
  name?: string | null;
  displayName?: string | null;
  description?: string | null;
  parentName?: string | null;
  isDbPermission?: boolean;
  parent?: PermissionDto;
  child?: PermissionDto[] | null;
}